requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
firebase-admin==6.5.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
resend==2.0.0
razorpay==1.4.1
httpx==0.27.0
pydantic==2.7.1
